requirements.txt
django==3.1.7
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-nested-inline==0.4.2
django-autoslug==1.9.8
psycopg2-binary==2.8.6
Pillow==8.2.0